+++ CHANGELOG.md
@@ -6,12 +6,14 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - Added support for enhanced aliases (#67)
 - Added support for per-alias image specification (#68)
+- Add bash completion support (#69)
 
 ### Changed
 - All ancillary files are bind-mounted via single temp dir
 - Hook scripts are moved to hooks/ subdirectory
 - User commands always executed via shell (#66)
 - Top-level directory mounted at same path in container (#70)
+- Alias names cannot contain spaces
 
 
 ## [1.7.0] - 2016-05-19